requirements.txt
python-dotenv
PyMuPDF
nltk
openai
#openai migrate

mysql-connector-python==8.4.0
python-dotenv==1.0.1
mysqlclient==2.2.4
pymysql==1.1.0
